requirements.txt
django==2.2.1
djangorestframework==3.9.2
gunicorn==19.9.0
whitenoise==4.1.2
dj-database-url==0.5.0
psycopg2==2.7.4